+++ NEWS.md
@@ -8,11 +8,14 @@
 
 * `linear_filter` gave totally wrong predictions due to a code error:  fixed.
 * `linear_filter` returned a matrix when NAs were present: fixed.
+* `fitted` now has an argument `labels` which allows to add the
+labels to the returned object.
 
 * In the testing procedures
     - testing skewed homogenous networks added.
     - testing validations added
     - testing symmetric calculations
+    - testing processing of labels
     
 * input testing for `tskrr` moved to its own function and is 
 also used by `impute_tskrr` now.